refactor: split EmitAcceptCaseParticipantOfferNode into accept_offer submodule (BTND-07-004)
After merging with main (which added RecordRecommendationRecommenderNode to
emit.py), the combined suggest_actor/emit.py exceeded the 500-line BTND-07-004
limit (523 lines in the PR-merge check). Move the owner-side Accept response
node EmitAcceptCaseParticipantOfferNode into its own accept_offer.py submodule
— a distinct semantic concern (Case Owner response) from the CaseActor-side
forwarding/notification emit nodes. Re-export via the suggest_actor package
__init__ (BTND-07-003) so importers are unaffected; actor_trigger_trees.py
imports it from the new module.
