+++ vultron/core/behaviors/case/nodes/suggest_actor/emit.py
@@ -27,9 +27,11 @@ Node classes:
 - :class:`EmitNoteDuplicateRecommendationToOwnerNode` — sends a
   ``Create(Note)`` + ``Add(Note, Case)`` to the Case Owner when a duplicate
   recommendation arrives (CM-16-008).
-- :class:`EmitAcceptCaseParticipantOfferNode` — Case Owner sends
-  ``Accept(Offer(CaseParticipant))`` back to the CaseActor (ADR-0026,
-  CM-16-006 owner-side response).
+
+The Case Owner owner-side Accept response
+(:class:`~vultron.core.behaviors.case.nodes.suggest_actor.accept_offer.EmitAcceptCaseParticipantOfferNode`)
+lives in the ``accept_offer`` submodule to keep this module under the
+BTND-07-004 line limit.
 """
 
 from typing import cast
@@ -396,70 +398,8 @@ class EmitNoteDuplicateRecommendationToOwnerNode(DataLayerAction):
             return Status.FAILURE
 
 
-class EmitAcceptCaseParticipantOfferNode(DataLayerAction):
-    """Case Owner sends Accept(Offer(CaseParticipant)) back to the CaseActor.
-
-    Triggered by the Case Owner after reviewing the Offer(CaseParticipant)
-    forwarded per ADR-0026 (CM-16-006 owner-side response).  The Accept is
-    addressed to the CaseActor so it can proceed with inviting the suggested
-    actor.
-    """
-
-    def __init__(
-        self,
-        cp_offer_id: str,
-        case_actor_id: str,
-        captured: dict | None = None,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.cp_offer_id = cp_offer_id
-        self.case_actor_id = case_actor_id
-        self._captured = captured
-
-    def _emit(self) -> tuple[str, dict]:
-        assert self.trigger_activity_factory is not None
-        assert self.actor_id is not None
-        return self.trigger_activity_factory.accept_case_participant_offer(
-            cp_offer_id=self.cp_offer_id,
-            actor=self.actor_id,
-            to=[self.case_actor_id],
-        )
-
-    def update(self) -> Status:
-        fail = self._require_datalayer_and_actor()
-        if fail is not None:
-            return fail
-        fail = self._require_factory()
-        if fail is not None:
-            self.logger.error(self.feedback_message)
-            return fail
-
-        try:
-            activity_id, activity_dict = self._emit()
-            cast(CaseOutboxPersistence, self.datalayer).record_outbox_item(
-                self.actor_id, activity_id  # type: ignore[arg-type]
-            )
-            if self._captured is not None:
-                self._captured["activity"] = activity_dict
-            self.logger.info(
-                "Actor '%s' accepted Offer(CaseParticipant) '%s' → CaseActor '%s'",
-                self.actor_id,
-                self.cp_offer_id,
-                self.case_actor_id,
-            )
-            return Status.SUCCESS
-        except Exception as e:
-            self.feedback_message = (
-                f"EmitAcceptCaseParticipantOffer failed: {e}"
-            )
-            self.logger.error(self.feedback_message)
-            return Status.FAILURE
-
-
 __all__ = [
     "EmitAcceptActorRecommendationNode",
-    "EmitAcceptCaseParticipantOfferNode",
     "EmitNoteDuplicateRecommendationToOwnerNode",
     "EmitOfferCaseParticipantToOwnerNode",
     "EmitRejectActorRecommendationNode",
